[ADDED] chose acount type
