+++ src/components/register/index.js
@@ -1,0 +1,4 @@
+import Client from "./Client";
+import AcountMan from "./AcountMan";
+
+export { Client, AcountMan };
